Pathing went to shit
Paths have been removed due to spaghetti code and weird bugs.

Orbitting planets is now encouraged, simply place an empty object where you want the center of rotation to be and play with the 'rotationAxis' and 'DegreesPerSecond' varaibles. Be sure to have the planet be a child of this empty object.